Refactor out server
