+++ package.json
@@ -44,6 +44,7 @@
     "@babel/core": "7.0.0-beta.46",
     "@babel/plugin-proposal-class-properties": "7.0.0-beta.46",
     "@babel/plugin-proposal-decorators": "7.0.0-beta.46",
+    "@babel/plugin-proposal-export-default-from": "^7.0.0-beta.52",
     "@babel/plugin-transform-runtime": "7.0.0-beta.46",
     "@babel/polyfill": "7.0.0-beta.46",
     "@babel/preset-env": "7.0.0-beta.46",
