+++ webpack.config.js
@@ -69,6 +69,7 @@ module.exports = (env, argv) => {
                   regenerator: true,
                 },
               ],
+              ['@babel/plugin-proposal-export-default-from'],
             ],
           },
         },
